Check if ~/.ssh/config exists before grepping it.

+++ .bash_profile
@@ -9,7 +9,7 @@ done
 shopt -s nocaseglob
 
 # Add tab completion for SSH hostnames based on ~/.ssh/config, ignoring wildcards
-complete -o "default" -o "nospace" -W "$(grep "^Host" ~/.ssh/config | grep -v "[?*]" | cut -d " " -f2)" scp sftp ssh
+[ -e "$HOME/.ssh/config" ] && complete -o "default" -o "nospace" -W "$(grep "^Host" ~/.ssh/config | grep -v "[?*]" | cut -d " " -f2)" scp sftp ssh
 
 # Add tab completion for `defaults read|write NSGlobalDomain`
 complete -W "NSGlobalDomain" defaults
